move logger types to separate file
