+++ packages/types/src/index.ts
@@ -3,3 +3,4 @@ export * from './utility-types.js';
 export * from './browser.js';
 export * from './http-client.js';
 export * from './session.js';
+export * from './logger.js';
